Revert "[web] Switch to doubles (#37336)" (#37609)
This reverts commit 6976132e5c794423693ac40ec4c705a9c736285c.

+++ lib/web_ui/lib/src/engine/html/surface_stats.dart
@@ -97,7 +97,7 @@ class DebugSurfaceStats {
 DomCanvasRenderingContext2D? _debugSurfaceStatsOverlayCtx;
 
 void debugRepaintSurfaceStatsOverlay(PersistedScene scene) {
-  final int overlayWidth = domWindow.innerWidth!.toInt();
+  final int overlayWidth = domWindow.innerWidth!;
   const int rowHeight = 30;
   const int rowCount = 4;
   const int overlayHeight = rowHeight * rowCount;
@@ -296,7 +296,7 @@ void debugPrintSurfaceStats(PersistedScene scene, int frameNumber) {
     final int pixelCount = canvasElements
         .cast<DomCanvasElement>()
         .map<int>((DomCanvasElement e) {
-      final int pixels = (e.width! * e.height!).toInt();
+      final int pixels = e.width! * e.height!;
       canvasInfo.writeln('    - ${e.width!} x ${e.height!} = $pixels pixels');
       return pixels;
     }).fold(0, (int total, int pixels) => total + pixels);
